+++ srcs/fractol.c
@@ -6,16 +6,16 @@
 /*   By: fmadura <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/09 16:57:44 by fmadura           #+#    #+#             */
-/*   Updated: 2018/07/22 18:54:23 by fmadura          ###   ########.fr       */
+/*   Updated: 2018/07/26 17:59:16 by fmadura          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fractol.h"
 
-t_env		*fractol_iter(t_env *env, int (algo)(t_env *, int, int, t_complex))
+t_env		*fractol_iter(t_env *env, int (algo)(t_env *, double, double, t_complex))
 {
-	int			x;
-	int			y;
+	double		x;
+	double		y;
 	int			pos;
 	t_complex	c;
